Cleaned up Cassandra functions, beginning to insert into Archivist
