+++ mainproject/cluster/src/main/scala/com/raphtory/core/model/graphentities/Vertex.scala
@@ -1,6 +1,6 @@
 package com.raphtory.core.model.graphentities
 
-import com.raphtory.core.actors.partitionmanager.Archivist.{VertexHistoryPoint, VertexPropertyPoint}
+import com.raphtory.core.storage.{VertexHistoryPoint, VertexPropertyPoint}
 
 import scala.collection.mutable
 import scala.collection.parallel.ParSet
